factor out some classes to be used in analysis as well
